requirements.txt
Django==5.1.2
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
django-cors-headers==4.5.0
async-timeout==4.0.3
django-redis==5.4.0
redis==5.1.1
channels==4.1.0
django-js-asset==2.2.0
django-ckeditor==6.7.1